Moves utilities to geometry module
Refactors the `find_extreme_coordinates` function and general utilities from the `delaunay_core` module to the more appropriate `geometry` module.

This change improves code organization and clarifies module responsibilities. It also removes the dedicated utilities.rs file from delaunay_core since its functions are now in the geometry module.
