+++ src/geometry/predicates.rs
@@ -11,7 +11,8 @@ use nalgebra as na;
 use num_traits::Float;
 use peroxide::fuga::{LinearAlgebra, MatrixTrait, anyhow, zeros};
 use serde::{Serialize, de::DeserializeOwned};
-use std::{hash::Hash, iter::Sum};
+use std::{cmp::Ordering, collections::HashMap, hash::Hash, iter::Sum};
+use uuid::Uuid;
 
 /// Calculate the circumcenter of a set of vertices forming a simplex.
 ///
@@ -459,6 +460,94 @@ where
     Ok(det < 0.0)
 }
 
+/// Find the extreme coordinates (minimum or maximum) across all vertices in a `HashMap`.
+///
+/// This function takes a `HashMap` of vertices and returns the minimum or maximum
+/// coordinates based on the specified ordering.
+///
+/// # Arguments
+///
+/// * `vertices` - A `HashMap` containing Vertex objects, where the key is a
+///   Uuid and the value is a Vertex.
+/// * `ordering` - The ordering parameter is of type Ordering and is used to
+///   specify whether the function should find the minimum or maximum
+///   coordinates. Ordering is an enum with three possible values: `Less`,
+///   `Equal`, and `Greater`.
+///
+/// # Returns
+///
+/// An array of type `T` with length `D` containing the minimum or maximum
+/// coordinate for each dimension.
+///
+/// # Panics
+///
+/// This function should not panic under normal circumstances as it handles
+/// the empty vertices case by returning default coordinates. However, it uses
+/// `.unwrap()` internally which could theoretically panic if the `HashMap`
+/// iterator behavior changes unexpectedly.
+///
+/// # Example
+///
+/// ```
+/// use d_delaunay::geometry::predicates::find_extreme_coordinates;
+/// use d_delaunay::delaunay_core::vertex::Vertex;
+/// use d_delaunay::geometry::point::Point;
+/// use std::collections::HashMap;
+/// use std::cmp::Ordering;
+/// let points = vec![
+///     Point::new([-1.0, 2.0, 3.0]),
+///     Point::new([4.0, -5.0, 6.0]),
+///     Point::new([7.0, 8.0, -9.0]),
+/// ];
+/// let vertices: Vec<Vertex<f64, Option<()>, 3>> = Vertex::from_points(points);
+/// let hashmap = Vertex::into_hashmap(vertices);
+/// let min_coords = find_extreme_coordinates(&hashmap, Ordering::Less);
+/// # use approx::assert_relative_eq;
+/// assert_relative_eq!(min_coords.as_slice(), [-1.0, -5.0, -9.0].as_slice(), epsilon = 1e-9);
+/// ```
+#[must_use]
+pub fn find_extreme_coordinates<T, U, const D: usize, S: ::std::hash::BuildHasher>(
+    vertices: &HashMap<Uuid, Vertex<T, U, D>, S>,
+    ordering: Ordering,
+) -> [T; D]
+where
+    T: Default + OrderedEq + Float,
+    U: Clone + Copy + Eq + Hash + Ord + PartialEq + PartialOrd,
+    [T; D]: Default + DeserializeOwned + Serialize + Sized,
+{
+    if vertices.is_empty() {
+        return [Default::default(); D];
+    }
+
+    // Initialize with the first vertex's coordinates using implicit conversion
+    let mut extreme_coords: [T; D] = vertices.values().next().unwrap().into();
+
+    // Compare with remaining vertices
+    for vertex in vertices.values().skip(1) {
+        let vertex_coords: [T; D] = vertex.into();
+        for (i, coord) in vertex_coords.iter().enumerate() {
+            match ordering {
+                Ordering::Less => {
+                    if *coord < extreme_coords[i] {
+                        extreme_coords[i] = *coord;
+                    }
+                }
+                Ordering::Greater => {
+                    if *coord > extreme_coords[i] {
+                        extreme_coords[i] = *coord;
+                    }
+                }
+                Ordering::Equal => {
+                    // For Equal ordering, return the first vertex's coordinates
+                    // This behavior maintains backward compatibility
+                }
+            }
+        }
+    }
+
+    extreme_coords
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -840,5 +929,227 @@ mod tests {
 
         assert!(circumsphere_far.is_ok());
         assert!(determinant_far.is_ok());
+    }
+
+    #[test]
+    fn predicates_find_min_coordinate() {
+        let points = vec![
+            Point::new([-1.0, 2.0, 3.0]),
+            Point::new([4.0, -5.0, 6.0]),
+            Point::new([7.0, 8.0, -9.0]),
+        ];
+        let vertices: Vec<crate::delaunay_core::vertex::Vertex<f64, Option<()>, 3>> = crate::delaunay_core::vertex::Vertex::from_points(points);
+        let hashmap = crate::delaunay_core::vertex::Vertex::into_hashmap(vertices);
+        let min_coords = find_extreme_coordinates(&hashmap, Ordering::Less);
+
+        assert_relative_eq!(
+            min_coords.as_slice(),
+            [-1.0, -5.0, -9.0].as_slice(),
+            epsilon = 1e-9
+        );
+
+        // Human readable output for cargo test -- --nocapture
+        println!("min_coords = {min_coords:?}");
+    }
+
+    #[test]
+    fn predicates_find_max_coordinate() {
+        let points = vec![
+            Point::new([-1.0, 2.0, 3.0]),
+            Point::new([4.0, -5.0, 6.0]),
+            Point::new([7.0, 8.0, -9.0]),
+        ];
+        let vertices: Vec<crate::delaunay_core::vertex::Vertex<f64, Option<()>, 3>> = crate::delaunay_core::vertex::Vertex::from_points(points);
+        let hashmap = crate::delaunay_core::vertex::Vertex::into_hashmap(vertices);
+        let max_coords = find_extreme_coordinates(&hashmap, Ordering::Greater);
+
+        assert_relative_eq!(
+            max_coords.as_slice(),
+            [7.0, 8.0, 6.0].as_slice(),
+            epsilon = 1e-9
+        );
+
+        // Human readable output for cargo test -- --nocapture
+        println!("max_coords = {max_coords:?}");
+    }
+
+    #[test]
+    fn predicates_find_extreme_coordinates_empty() {
+        let empty_hashmap: HashMap<Uuid, crate::delaunay_core::vertex::Vertex<f64, Option<()>, 3>> = HashMap::new();
+        let min_coords = find_extreme_coordinates(&empty_hashmap, Ordering::Less);
+        let max_coords = find_extreme_coordinates(&empty_hashmap, Ordering::Greater);
+
+        // With empty hashmap, should return default values [0.0, 0.0, 0.0]
+        assert_relative_eq!(
+            min_coords.as_slice(),
+            [0.0, 0.0, 0.0].as_slice(),
+            epsilon = 1e-9
+        );
+        assert_relative_eq!(
+            max_coords.as_slice(),
+            [0.0, 0.0, 0.0].as_slice(),
+            epsilon = 1e-9
+        );
+    }
+
+    #[test]
+    fn predicates_find_extreme_coordinates_single_point() {
+        let points = vec![Point::new([5.0, -3.0, 7.0])];
+        let vertices: Vec<crate::delaunay_core::vertex::Vertex<f64, Option<()>, 3>> = crate::delaunay_core::vertex::Vertex::from_points(points);
+        let hashmap = crate::delaunay_core::vertex::Vertex::into_hashmap(vertices);
+
+        let min_coords = find_extreme_coordinates(&hashmap, Ordering::Less);
+        let max_coords = find_extreme_coordinates(&hashmap, Ordering::Greater);
+
+        // With single point, min and max should be the same
+        assert_relative_eq!(
+            min_coords.as_slice(),
+            [5.0, -3.0, 7.0].as_slice(),
+            epsilon = 1e-9
+        );
+        assert_relative_eq!(
+            max_coords.as_slice(),
+            [5.0, -3.0, 7.0].as_slice(),
+            epsilon = 1e-9
+        );
+    }
+
+    #[test]
+    fn predicates_find_extreme_coordinates_equal_ordering() {
+        let points = vec![Point::new([1.0, 2.0, 3.0]), Point::new([4.0, 5.0, 6.0])];
+        let vertices: Vec<crate::delaunay_core::vertex::Vertex<f64, Option<()>, 3>> = crate::delaunay_core::vertex::Vertex::from_points(points);
+        let hashmap = crate::delaunay_core::vertex::Vertex::into_hashmap(vertices);
+
+        // Using Ordering::Equal should return the first vertex's coordinates unchanged
+        let coords = find_extreme_coordinates(&hashmap, Ordering::Equal);
+        // The first vertex in the iteration (order is not guaranteed in HashMap)
+        // but the result should be one of the input coordinates
+        let matches_first = approx::relative_eq!(
+            coords.as_slice(),
+            [1.0, 2.0, 3.0].as_slice(),
+            epsilon = 1e-9
+        );
+        let matches_second = approx::relative_eq!(
+            coords.as_slice(),
+            [4.0, 5.0, 6.0].as_slice(),
+            epsilon = 1e-9
+        );
+        assert!(matches_first || matches_second);
+    }
+
+    #[test]
+    fn predicates_find_extreme_coordinates_2d() {
+        let points = vec![
+            Point::new([1.0, 4.0]),
+            Point::new([3.0, 2.0]),
+            Point::new([2.0, 5.0]),
+        ];
+        let vertices: Vec<crate::delaunay_core::vertex::Vertex<f64, Option<()>, 2>> = crate::delaunay_core::vertex::Vertex::from_points(points);
+        let hashmap = crate::delaunay_core::vertex::Vertex::into_hashmap(vertices);
+
+        let min_coords = find_extreme_coordinates(&hashmap, Ordering::Less);
+        let max_coords = find_extreme_coordinates(&hashmap, Ordering::Greater);
+
+        assert_relative_eq!(min_coords.as_slice(), [1.0, 2.0].as_slice(), epsilon = 1e-9);
+        assert_relative_eq!(max_coords.as_slice(), [3.0, 5.0].as_slice(), epsilon = 1e-9);
+    }
+
+    #[test]
+    fn predicates_find_extreme_coordinates_1d() {
+        let points = vec![Point::new([10.0]), Point::new([-5.0]), Point::new([3.0])];
+        let vertices: Vec<crate::delaunay_core::vertex::Vertex<f64, Option<()>, 1>> = crate::delaunay_core::vertex::Vertex::from_points(points);
+        let hashmap = crate::delaunay_core::vertex::Vertex::into_hashmap(vertices);
+
+        let min_coords = find_extreme_coordinates(&hashmap, Ordering::Less);
+        let max_coords = find_extreme_coordinates(&hashmap, Ordering::Greater);
+
+        assert_relative_eq!(min_coords.as_slice(), [-5.0].as_slice(), epsilon = 1e-9);
+        assert_relative_eq!(max_coords.as_slice(), [10.0].as_slice(), epsilon = 1e-9);
+    }
+
+    #[test]
+    fn predicates_find_extreme_coordinates_with_typed_data() {
+        let points = vec![
+            Point::new([1.0, 2.0, 3.0]),
+            Point::new([4.0, -1.0, 2.0]),
+            Point::new([-2.0, 5.0, 1.0]),
+        ];
+        let vertices: Vec<crate::delaunay_core::vertex::Vertex<f64, i32, 3>> = points
+            .into_iter()
+            .enumerate()
+            .map(|(i, point)| {
+                VertexBuilder::default()
+                    .point(point)
+                    .data(i32::try_from(i).unwrap())
+                    .build()
+                    .unwrap()
+            })
+            .collect();
+        let hashmap = crate::delaunay_core::vertex::Vertex::into_hashmap(vertices);
+
+        let min_coords = find_extreme_coordinates(&hashmap, Ordering::Less);
+        let max_coords = find_extreme_coordinates(&hashmap, Ordering::Greater);
+
+        assert_relative_eq!(
+            min_coords.as_slice(),
+            [-2.0, -1.0, 1.0].as_slice(),
+            epsilon = 1e-9
+        );
+        assert_relative_eq!(
+            max_coords.as_slice(),
+            [4.0, 5.0, 3.0].as_slice(),
+            epsilon = 1e-9
+        );
+    }
+
+    #[test]
+    fn predicates_find_extreme_coordinates_identical_points() {
+        let points = vec![
+            Point::new([2.0, 3.0, 4.0]),
+            Point::new([2.0, 3.0, 4.0]),
+            Point::new([2.0, 3.0, 4.0]),
+        ];
+        let vertices: Vec<crate::delaunay_core::vertex::Vertex<f64, Option<()>, 3>> = crate::delaunay_core::vertex::Vertex::from_points(points);
+        let hashmap = crate::delaunay_core::vertex::Vertex::into_hashmap(vertices);
+
+        let min_coords = find_extreme_coordinates(&hashmap, Ordering::Less);
+        let max_coords = find_extreme_coordinates(&hashmap, Ordering::Greater);
+
+        // All points are identical, so min and max should be the same
+        assert_relative_eq!(
+            min_coords.as_slice(),
+            [2.0, 3.0, 4.0].as_slice(),
+            epsilon = 1e-9
+        );
+        assert_relative_eq!(
+            max_coords.as_slice(),
+            [2.0, 3.0, 4.0].as_slice(),
+            epsilon = 1e-9
+        );
+    }
+
+    #[test]
+    fn predicates_find_extreme_coordinates_large_numbers() {
+        let points = vec![
+            Point::new([1e6, -1e6, 1e12]),
+            Point::new([-1e9, 1e3, -1e15]),
+            Point::new([1e15, 1e9, 1e6]),
+        ];
+        let vertices: Vec<crate::delaunay_core::vertex::Vertex<f64, Option<()>, 3>> = crate::delaunay_core::vertex::Vertex::from_points(points);
+        let hashmap = crate::delaunay_core::vertex::Vertex::into_hashmap(vertices);
+
+        let min_coords = find_extreme_coordinates(&hashmap, Ordering::Less);
+        let max_coords = find_extreme_coordinates(&hashmap, Ordering::Greater);
+
+        assert_relative_eq!(
+            min_coords.as_slice(),
+            [-1e9, -1e6, -1e15].as_slice(),
+            epsilon = 1e-9
+        );
+        assert_relative_eq!(
+            max_coords.as_slice(),
+            [1e15, 1e9, 1e12].as_slice(),
+            epsilon = 1e-9
+        );
     }
 }
